requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
djangorestframework==3.15.2
idna==3.7
pusher-push-notifications==2.0.1
PyJWT==2.9.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.2
whitenoise==6.7.0
